Remove the dependency on a running service when reading params
The previous version used the service definition to look up subnet
and VPC ids.  This means that there needs to be a running service,
creating a chicken-and-egg situation: to deploy the service, you need
the params, to get the params, you need a deployed service.

Every cluster created by `ecs-cli compose up` has a corresponding
CloudFormation stack that tracks all the resources in the cluster.
This newer version inspects the CloudFormation resources directly.