Explicitly set the region when making ecs-cli calls
The ecs-cli uses a different order of precedence for setting the region than
the aws-cli which can result in errors and/or unexpected results in scripts.

In particular, ecs-cli will use the region set in the default ecs-cli profile
before using the region set in the environment (AWS_DEFAULT_REGION).

For the full order of precedence see the "order of precedence for region" here:
https://docs.aws.amazon.com/AmazonECS/latest/userguide/ECS_CLI_Configuration.html