Move all environment variable setup into compose file
As we try to consolidate configs, it's been difficult to figure out where
a particular value is being set based on multiple instances of enviroment
settings and the corresponding order of precedence.

As a first step, this PR brings all the enviroment into docker-compose.yml
so that we have a single file to manage.  Once we have working local and
ECS settings, we can revisit which settings it makes sense to manage via
separate config files.