Add additonal configuration necessary to run in production
* Environment Variables
A number of initialization and setup tasks like database migrations and Solr
initialization are run outside the web server context and therefore need the
RAILS_ENV environment variable to be explicitly set to "production".

* Database
The management application previously had non-default valued configured for the
production database username and password.  The multi-database configuration we
currently use for the blacklight and management apps does not support setting
distict usernames and passwords on separate databases, so a new version of the
management app was released that uses consistent database credentials in the
production database config settings.