Add a deployment check for the RAILS_MASTER_KEY
The RAILS_MASTER_KEY is used to decrypt encoded credentials checked into
the yul-dc-blacklight and yul-dc-management repos.  We have agreed to use
the same credentials for both repos and to share the encryption key out-of-band
with team members.

The key must be a 32-character hex string representing a 16-byte encoding key.
The key must be set in the shell environment of the machine running the `deploy-full.sh`
script.

For more details see https://edgeguides.rubyonrails.org/security.html#environmental-security